Allow parallel requests over the same SSH session
Introducing a simple mutex mechanism to virtually lock the
channel in the current SSH session. During this time, all
subsequent requests will be waiting to be served.
If there are any requests waiting more than the timeout
specified during open(), CommandTimeoutException is raised.
Request from other SSH sessions are ot affected by this.
This should not affect at all the existing behaviour,
but it will widen the scope of the library.